+++ site/dat/docs/Changelog.md
@@ -5,6 +5,7 @@
  - support variables in time fields for JMeter
  - support POST body in application/json format
  - add `-locations` command alias to dump available locations
+ - ability to set JMeter log verbosity with `write-xml-jtl`
  - add concurrency-driven load to PBench
 
 ## 1.0.0